Update __init__.py
Fixing View#wait_until_displayed/wait_until_not_displayed

It was using `self.driver` as an argument to `WebDriverWait` which fails when initializing the `View` with a callable. This is because in that init scenario the driver is derived from `self.root` which will throw `NoSuchElementException` if its not on the page at the time.

The driver is only being passed to satisfy `WebDriverWait` anyways. Its not actually used.